вернул double чтобы соответствовать условиям задачи

+++ src/main/java/site/javadev/lesson_03/Task7.java
@@ -17,9 +17,9 @@ public class Task7 {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.println("Введите время в минутах: ");
-        int t = scanner.nextInt();  // Используем int для вещественного числа
+        double t = scanner.nextDouble();  // Используем double для вещественного числа
 
-        int remainder = t % 5;
+        double remainder = t % 5;
 
         if(remainder <3){
             System.out.println("Зеленый сигнал");
@@ -29,9 +29,6 @@ public class Task7 {
         }else {
             System.out.println("Красный сигнал");  // в остальных случаях остаток будет 4
         }
-
-
-
     }
 
 }
